Refactor: Split AcademicCalendar component
Splits the AcademicCalendar component into smaller, more manageable files for improved code organization and maintainability. No functional changes.
